adding BaseLocationFragment to handle common location stuff
