+++ Mobile/Android/app/src/main/java/com/utn/frba/desarrollomobile/hunter/ui/fragment/AppFragment.kt
@@ -23,7 +23,7 @@ class AppFragment : Fragment(R.layout.fragment_app) {
         }
 
         compassButton.setOnClickListener {
-            showFragment(DummyFragment(), true)
+            showFragment(MapFragment(), true)
         }
     }
 }
